image uploaded
I guess I'm working on the end product first, little by little the features will be added to the codebase. Right now the user picks an image, and the script takes that image, makes a copy, and moves it to the uploads folder as a finished project. we still to rename the image, and resize it. and at the end maybe the software can edit the images in batches.