Google Sheets Integration
There is a new module that will edit the Google Sheets and configure the app to submit data to the sheets. Right now, I need to make 3 manual updates automatic, based on time of day, or by the push of a button.  I still need to update the roadmap, but this is where we are.